Revert "now go through door without pressing space; just both players need to be colliding. Repositioned pressure plate so it doesn't bug out with uneven surface when sinking"
This reverts commit 064da572e5343738983587f7b893563c927ec79c.

+++ Assets/Scripts/PressurePlateRotation/PressurePlate.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 public class PressurePlate : MonoBehaviour
 {
     public PressurePlateController parent;
-    public float distToSink = 0.2f;
+    public float distToSink = (float)0.2;
     public int animationTimeInFrames = 10;
 
     private List<GameObject> localColliders;
